+++ js/paramviewer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/paramviewer
 
+## 0.7.4
+
+### Dependency updates
+
+- @gradio/atoms@0.14.1
+- @gradio/statustracker@0.10.6
+
 ## 0.7.3
 
 ### Dependency updates