+++ js/slider/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/slider
 
+## 0.6.4
+
+### Dependency updates
+
+- @gradio/atoms@0.14.1
+- @gradio/statustracker@0.10.6
+
 ## 0.6.3
 
 ### Dependency updates